refactor: move return spy and stub to plugin
