+++ src/genericFunction/index.ts
@@ -1,14 +1,16 @@
 import { spyFunction } from './getSpyFunction'
-import { stubFunction } from './getStubFunction';
+import { stubFunction } from './getStubFunction'
 
-export const genericFunction = {
-  canHandleSubject(subject) {
-    return typeof subject === 'function'
-  },
-  getSpy({ resolve, store }, subject) {
-    return spyFunction({ resolve, store }, subject)
-  },
-  getStub({ resolve, store }, subject, id) {
-    return stubFunction({ resolve, store }, subject, id)
-  }
+let komondor
+
+export function activate(k) {
+  komondor = k
+}
+
+export function getSpy({ resolve, store }, subject) {
+  return spyFunction({ komondor, resolve, store }, subject)
+}
+
+export function getStub({ resolve, store }, subject, id) {
+  return stubFunction({ komondor, resolve, store }, subject, id)
 }
